refactor: remove model provider resolution from infra layer
Move checkOrgCredits() and checkModelProviderConfigured() from the
generic run-service.ts to a new zero-preflight.ts module, called
exclusively from createZeroRun(). The generic `vm0 run` / SDK path
no longer auto-resolves model provider secrets — users must pass
API keys explicitly via --secrets.

Key changes:
- Remove modelProvider from unifiedRunRequestSchema (keep in zero schema)
- Remove --model-provider flag from vm0 run/continue/resume CLI commands
- Remove checkOrgCredits and checkModelProviderConfigured from createRunRecord
- Create zero-preflight.ts with both functions for Zero layer use
- Call preflight checks from createZeroRun() before createRunRecord()
- Export loadCompose from run-service for Zero preflight access
- Migrate credit-check and pre-run-checks tests to zero/__tests__/

Part of #7152 (Extract Zero Platform Layer from Generic Run Infrastructure)
Closes #7353
