+++ turbo/apps/cli/src/commands/run/__tests__/continue.test.ts
@@ -287,35 +287,6 @@ describe("run continue command", () => {
         }),
       );
     });
-
-    it("should pass model provider option to API", async () => {
-      let capturedBody: Record<string, unknown> | undefined;
-
-      server.use(
-        http.post(
-          "http://localhost:3000/api/agent/runs",
-          async ({ request }) => {
-            capturedBody = (await request.json()) as Record<string, unknown>;
-            return HttpResponse.json(defaultRunResponse, { status: 201 });
-          },
-        ),
-      );
-
-      await continueCommand.parseAsync([
-        "node",
-        "cli",
-        testSessionId,
-        "test prompt",
-        "--model-provider",
-        "anthropic-api-key",
-      ]);
-
-      expect(capturedBody).toEqual(
-        expect.objectContaining({
-          modelProvider: "anthropic-api-key",
-        }),
-      );
-    });
   });
 
   describe("session ID validation", () => {
