+++ turbo/apps/cli/src/commands/run/__tests__/resume.test.ts
@@ -294,35 +294,6 @@ describe("run resume command", () => {
         }),
       );
     });
-
-    it("should pass model provider option to API", async () => {
-      let capturedBody: Record<string, unknown> | undefined;
-
-      server.use(
-        http.post(
-          "http://localhost:3000/api/agent/runs",
-          async ({ request }) => {
-            capturedBody = (await request.json()) as Record<string, unknown>;
-            return HttpResponse.json(defaultRunResponse, { status: 201 });
-          },
-        ),
-      );
-
-      await resumeCommand.parseAsync([
-        "node",
-        "cli",
-        testCheckpointId,
-        "test prompt",
-        "--model-provider",
-        "anthropic-api-key",
-      ]);
-
-      expect(capturedBody).toEqual(
-        expect.objectContaining({
-          modelProvider: "anthropic-api-key",
-        }),
-      );
-    });
   });
 
   describe("checkpoint ID validation", () => {
