+++ turbo/apps/web/src/lib/run/__tests__/credit-check.test.ts
@@ -7,13 +7,9 @@ import {
 import {
   createTestCompose,
   findTestRunRecord,
-  findTestRunsByUserAndPrompt,
   findTestQueueEntry,
   markRunningRunsAsCompleted,
   setOrgCredits,
-  deleteOrgRow,
-  insertOrgDefaultModelProvider,
-  insertOrgMembersEntry,
 } from "../../../__tests__/api-test-helpers";
 import { reloadEnv } from "../../../env";
 import {
@@ -22,11 +18,10 @@ import {
   type CreateRunParams,
 } from "../run-service";
 import { drainOrgQueue, enqueueRun } from "../run-queue-service";
-import { isInsufficientCredits } from "../../errors";
 
 const context = testContext();
 
-describe("credit check", () => {
+describe("credit check — dequeue path", () => {
   let user: UserContext;
   let versionId: string;
 
@@ -46,160 +41,6 @@ describe("credit check", () => {
       ...overrides,
     };
   }
-
-  describe("createRun() path", () => {
-    it("should allow VM0 run when credits > 0", async () => {
-      await setOrgCredits(user.orgId, 100);
-
-      const result = await createRun(baseParams({ modelProvider: "vm0" }));
-
-      expect(result.status).toBe("pending");
-      expect(result.runId).toBeDefined();
-    });
-
-    it("should reject VM0 run when credits = 0", async () => {
-      await setOrgCredits(user.orgId, 0);
-
-      await expect(
-        createRun(baseParams({ modelProvider: "vm0" })),
-      ).rejects.toSatisfy(isInsufficientCredits);
-    });
-
-    it("should reject VM0 run when credits are negative", async () => {
-      await setOrgCredits(user.orgId, -500);
-
-      await expect(
-        createRun(baseParams({ modelProvider: "vm0" })),
-      ).rejects.toSatisfy(isInsufficientCredits);
-    });
-
-    it("should allow non-VM0 run when credits = 0", async () => {
-      await setOrgCredits(user.orgId, 0);
-
-      const result = await createRun(
-        baseParams({ modelProvider: "anthropic" }),
-      );
-
-      expect(result.status).toBe("pending");
-    });
-
-    it("should reject when org default is VM0 and credits = 0", async () => {
-      await setOrgCredits(user.orgId, 0);
-      await insertOrgDefaultModelProvider(user.orgId, "vm0");
-
-      await expect(createRun(baseParams())).rejects.toSatisfy(
-        isInsufficientCredits,
-      );
-    });
-
-    it("should allow when org default is non-VM0 and credits = 0", async () => {
-      await setOrgCredits(user.orgId, 0);
-      await insertOrgDefaultModelProvider(user.orgId, "anthropic-api-key");
-
-      const result = await createRun(baseParams());
-
-      expect(result.status).toBe("pending");
-    });
-
-    it("should allow when no org default provider and credits = 0", async () => {
-      await setOrgCredits(user.orgId, 0);
-
-      const result = await createRun(baseParams());
-
-      expect(result.status).toBe("pending");
-    });
-
-    it("should allow when org_metadata row is missing", async () => {
-      await deleteOrgRow(user.orgId);
-
-      const result = await createRun(baseParams({ modelProvider: "vm0" }));
-
-      expect(result.status).toBe("pending");
-    });
-
-    it("should not enqueue a rejected VM0 run", async () => {
-      await setOrgCredits(user.orgId, 0);
-
-      const prompt = "Rejected VM0 run - no enqueue";
-      await expect(
-        createRun(baseParams({ modelProvider: "vm0", prompt })),
-      ).rejects.toSatisfy(isInsufficientCredits);
-
-      // Verify no run record was created (credit check rejects before INSERT)
-      const runs = await findTestRunsByUserAndPrompt(user.userId, prompt);
-      expect(runs).toHaveLength(0);
-    });
-  });
-
-  describe("member credit cap enforcement", () => {
-    it("should reject VM0 run when creditEnabled is false", async () => {
-      await setOrgCredits(user.orgId, 10000);
-      await insertOrgDefaultModelProvider(user.orgId, "vm0");
-
-      // Disable member via credit cap
-      await insertOrgMembersEntry({
-        orgId: user.orgId,
-        userId: user.userId,
-        creditCap: 100,
-        creditEnabled: false,
-      });
-
-      await expect(
-        createRun(baseParams({ modelProvider: "vm0" })),
-      ).rejects.toSatisfy(isInsufficientCredits);
-    });
-
-    it("should allow non-VM0 run regardless of creditEnabled", async () => {
-      await setOrgCredits(user.orgId, 10000);
-
-      // Disable member via credit cap
-      await insertOrgMembersEntry({
-        orgId: user.orgId,
-        userId: user.userId,
-        creditCap: 100,
-        creditEnabled: false,
-      });
-
-      const result = await createRun(
-        baseParams({ modelProvider: "anthropic" }),
-      );
-      expect(result.status).toBe("pending");
-    });
-
-    it("should allow VM0 run when creditEnabled is true with cap set", async () => {
-      await setOrgCredits(user.orgId, 10000);
-      await insertOrgDefaultModelProvider(user.orgId, "vm0");
-
-      // Set cap but leave enabled
-      await insertOrgMembersEntry({
-        orgId: user.orgId,
-        userId: user.userId,
-        creditCap: 10000,
-        creditEnabled: true,
-      });
-
-      const result = await createRun(baseParams({ modelProvider: "vm0" }));
-      expect(result.status).toBe("pending");
-    });
-
-    it("should reject VM0 run when default provider is vm0 and creditEnabled is false", async () => {
-      await setOrgCredits(user.orgId, 10000);
-      await insertOrgDefaultModelProvider(user.orgId, "vm0");
-
-      // Disable member via credit cap
-      await insertOrgMembersEntry({
-        orgId: user.orgId,
-        userId: user.userId,
-        creditCap: 100,
-        creditEnabled: false,
-      });
-
-      // No explicit modelProvider — falls back to org default (vm0)
-      await expect(createRun(baseParams())).rejects.toSatisfy(
-        isInsufficientCredits,
-      );
-    });
-  });
 
   describe("dequeueNextAtomic() path", () => {
     it("should fail queued VM0 run when credits depleted at drain time", async () => {
