+++ turbo/packages/core/src/contracts/runs.ts
@@ -47,9 +47,6 @@ const unifiedRunRequestSchema = z.object({
   // Debug flag to force real Claude in mock environments (internal use only)
   debugNoMockClaude: z.boolean().optional(),
 
-  // Model provider for automatic secret injection
-  modelProvider: z.string().optional(),
-
   // Environment validation flag - when true, validates secrets/vars before running
   checkEnv: z.boolean().optional(),
